reuse changelogs if exists

+++ Kartverket.Geosynkronisering.Server/Provider_NetCore/DatasetStatus.cs
@@ -2,7 +2,7 @@
 {
     internal class DatasetStatus
     {
-        public long copy_transaction_token;
+        public int copy_transaction_token;
 
     }
 }
